remove company name
