+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.zand</groupId>
+    <groupId>com.greyson</groupId>
     <artifactId>Bookstore</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>Bookstore</name>
@@ -140,7 +140,7 @@
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>com.zand.bookstore.BookstoreApplication</mainClass>
+                    <mainClass>com.greyson.bookstore.BookstoreApplication</mainClass>
                     <layout>JAR</layout>
                 </configuration>
                 <executions>
