+++ src/main/java/com/greyson/bookstore/BookstoreApplication.java
@@ -1,0 +1,13 @@
+package com.greyson.bookstore;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BookstoreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BookstoreApplication.class, args);
+    }
+
+}
