+++ src/main/java/com/greyson/bookstore/common/Constants.java
@@ -1,0 +1,9 @@
+package com.greyson.bookstore.common;
+
+public class Constants {
+    public final static String NOT_ENOUGH_STOCK = " have not enough stock, please check it again";
+    public final static String INDICATOR_BOOK_START = "[";
+    public final static String INDICATOR_BOOK_END = "]";
+    public final static String SAVE_RESULT_MESSAGE = "Save %s book(s), Failed:%d";
+    public final static String ADD_FAILED_BOOK_ID = "Failed Book Id:";
+}
