+++ src/main/java/com/greyson/bookstore/config/SwaggerConfig.java
@@ -1,0 +1,21 @@
+package com.greyson.bookstore.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+    /**
+     * Define basic information for api page displaying
+     * @return
+     */
+    @Bean
+    public OpenAPI springShopOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("Book Store")
+                        .description("Swagger page for Greyson's book store")
+                        .version("v1"));
+    }
+}
