+++ src/main/java/com/greyson/bookstore/exception/BookStockException.java
@@ -1,0 +1,7 @@
+package com.greyson.bookstore.exception;
+
+public class BookStockException extends Exception{
+    public BookStockException(String message) {
+        super(message);
+    }
+}
