+++ src/main/java/com/greyson/bookstore/exception/ExceptionHandlerAdvice.java
@@ -1,0 +1,25 @@
+package com.greyson.bookstore.exception;
+
+import com.greyson.bookstore.common.CommonResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@ResponseBody
+@Slf4j
+@RestControllerAdvice
+public class ExceptionHandlerAdvice {
+    @ExceptionHandler(BookStockException.class)
+    public CommonResponse<Object> handleValidationExceptions(BookStockException ex) {
+        log.error("[handleValidationExceptions]", ex);
+        return CommonResponse.response(HttpStatus.BAD_REQUEST.value(), ex.getMessage(), null);
+    }
+
+    @ExceptionHandler(value = Exception.class)
+    public CommonResponse<?> defaultExceptionHandler(Throwable ex) {
+        log.error("[defaultExceptionHandler]", ex);
+        return CommonResponse.response(HttpStatus.INTERNAL_SERVER_ERROR.value(), HttpStatus.INTERNAL_SERVER_ERROR.toString(), null);
+    }
+}
