--- src/main/java/com/zand/bookstore/BookstoreApplication.java
@@ -1,13 +1,0 @@
-package com.zand.bookstore;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BookstoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BookstoreApplication.class, args);
-    }
-
-}
