--- src/main/java/com/zand/bookstore/config/SwaggerConfig.java
@@ -1,21 +1,0 @@
-package com.zand.bookstore.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    /**
-     * Define basic information for api page displaying
-     * @return
-     */
-    @Bean
-    public OpenAPI springShopOpenAPI() {
-        return new OpenAPI()
-                .info(new Info().title("Zand Book Store")
-                        .description("Swagger page for Zand booke store")
-                        .version("v1"));
-    }
-}
