--- src/main/java/com/zand/bookstore/exception/BookStockException.java
@@ -1,7 +1,0 @@
-package com.zand.bookstore.exception;
-
-public class BookStockException extends Exception{
-    public BookStockException(String message) {
-        super(message);
-    }
-}
