+++ src/test/java/com/greyson/bookstore/BookstoreApplicationTests.java
@@ -1,0 +1,14 @@
+package com.greyson.bookstore;
+
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BookstoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
