--- src/test/java/com/zand/bookstore/BookstoreApplicationTests.java
@@ -1,14 +1,0 @@
-package com.zand.bookstore;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BookstoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
